terraform/module/function_hoteles/requirements.txt
httpx==0.27.0
google-cloud-bigquery==3.17.2
google-cloud-logging==3.9.0
functions-framework==3.5.0
